flatten ws and curl responses
