+++ packages/insomnia/src/ui/components/websockets/realtime-response-pane.tsx
@@ -4,10 +4,12 @@ import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
 import { getSetCookieHeaders } from '../../../common/misc';
+import { CurlEvent } from '../../../main/network/curl';
 import { ResponseTimelineEntry } from '../../../main/network/libcurl-promise';
 import { WebSocketEvent } from '../../../main/network/websocket';
+import { Response } from '../../../models/response';
 import { WebSocketResponse } from '../../../models/websocket-response';
-import { useWebSocketConnectionEvents } from '../../context/websocket-client/use-ws-connection-events';
+import { useCurlConnectionEvents } from '../../context/websocket-client/use-ws-connection-events';
 import { selectActiveResponse } from '../../redux/selectors';
 import { PanelContainer, TabItem, Tabs } from '../base/tabs';
 import { ResponseHistoryDropdown } from '../dropdowns/response-history-dropdown';
@@ -82,11 +84,11 @@ const PaddedButton = styled('button')({
   padding: 'var(--padding-sm)',
 });
 
-export const WebSocketResponsePane: FC<{ requestId: string }> =
+export const RealtimeResponsePane: FC<{ requestId: string }> =
   ({
     requestId,
   }) => {
-    const response = useSelector(selectActiveResponse) as WebSocketResponse | null;
+    const response = useSelector(selectActiveResponse) as WebSocketResponse | Response | null;
     if (!response) {
       return (
         <Pane type="response">
@@ -95,33 +97,33 @@ export const WebSocketResponsePane: FC<{ requestId: string }> =
             icon={<i className="fa fa-paper-plane" />}
             documentationLinks={[
               {
-                title: 'Introduction to WebSockets in Insomnia',
+                title: 'Introduction to Curls in Insomnia',
                 url: 'https://docs.insomnia.rest/insomnia/requests',
               },
             ]}
-            title="Enter a URL and connect to a WebSocket server to start sending data"
+            title="Enter a URL and connect to a Curl server to start sending data"
             secondaryAction="Select a payload type to send data to the connection"
           />
         </Pane>
       );
     }
-    return <WebSocketActiveResponsePane requestId={requestId} response={response} />;
+    return <RealtimeActiveResponsePane requestId={requestId} response={response} />;
   };
 
-const WebSocketActiveResponsePane: FC<{ requestId: string; response: WebSocketResponse }> = ({
+const RealtimeActiveResponsePane: FC<{ requestId: string; response: WebSocketResponse | Response }> = ({
   requestId,
   response,
 }) => {
-  const [selectedEvent, setSelectedEvent] = useState<WebSocketEvent | null>(null);
+  const [selectedEvent, setSelectedEvent] = useState<CurlEvent | WebSocketEvent | null>(null);
   const [timeline, setTimeline] = useState<ResponseTimelineEntry[]>([]);
 
   const searchInputRef = useRef<HTMLInputElement>(null);
   const [clearEventsBefore, setClearEventsBefore] = useState<number | null>(null);
   const [searchQuery, setSearchQuery] = useState('');
-  const [eventType, setEventType] = useState<WebSocketEvent['type']>();
-  const allEvents = useWebSocketConnectionEvents({ responseId: response._id });
-  const handleSelection = (event: WebSocketEvent) => {
-    setSelectedEvent((selected: WebSocketEvent | null) => selected?._id === event._id ? null : event);
+  const [eventType, setEventType] = useState<CurlEvent['type']>();
+  const allEvents = useCurlConnectionEvents({ responseId: response._id });
+  const handleSelection = (event: CurlEvent | WebSocketEvent) => {
+    setSelectedEvent((selected: CurlEvent | WebSocketEvent | null) => selected?._id === event._id ? null : event);
   };
 
   const events = allEvents.filter(event => {
@@ -191,7 +193,7 @@ const WebSocketActiveResponsePane: FC<{ requestId: string; response: WebSocketRe
           className="tall pane__header__right"
         />
       </PaneHeader>
-      <Tabs aria-label="Websocket response pane tabs">
+      <Tabs aria-label="Curl response pane tabs">
         <TabItem key="events" title="Events">
           <PaneBodyContent>
             {response.error ? <ResponseErrorViewer url={response.url} error={response.error} />
@@ -205,7 +207,7 @@ const WebSocketActiveResponsePane: FC<{ requestId: string; response: WebSocketRe
                       gap: 'var(--padding-sm)',
                     }}
                   >
-                    <select onChange={e => setEventType(e.currentTarget.value as WebSocketEvent['type'])}>
+                    <select onChange={e => setEventType(e.currentTarget.value as CurlEvent['type'])}>
                       <option value="">All</option>
                       <option value="message">Message</option>
                       <option value="open">Open</option>
